repl: factor the REPL into a separate package
...so that users can build REPLs for their dialects.

Also, add support for interrupting the REPL with Control-C.
